fix: 랭킹 컴포넌트 수정, 닉네임 검색 컴포넌트 분리
랭킹 컴포넌트 서버 컴포넌트로 변경
닉네임 검색 컴포넌트 분리
